组件化代码结构

+++ app/src/main/java/com/example/weather/Ui/Place/PlaceViewModel/CityWeatherViewModel.java
@@ -1,14 +1,17 @@
 package com.example.weather.Ui.Place.PlaceViewModel;
 
 import android.app.Activity;
-import android.util.Log;
 
+import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
 import com.example.weather.Logic.WeatherDataInquireTool;
-import com.example.weather.Logic.model.TwentyFourHourWeatherData;
-import com.example.weather.Logic.netWorkUtil.HourlyWeatherData;
-import com.example.weather.Logic.netWorkUtil.HourlyWeatherUtility;
+import com.example.weather.Logic.model.TwentyFourHourWeatherDataModel;
+import com.example.weather.Logic.model.TwentyFourHourWeatherDatabase;
+import com.example.weather.Logic.netWorkUtil.LocationAndCity.HourlyWeatherData;
+import com.example.weather.Logic.netWorkUtil.LocationAndCity.HourlyWeatherUtility;
+import com.example.weather.Logic.netWorkUtil.LocationAndCity.SevenDayWeatherData;
+import com.example.weather.Logic.netWorkUtil.LocationAndCity.SevenDayWeatherUtility;
 import com.example.weather.TestTool.LogUtil;
 
 import java.io.IOException;
@@ -23,14 +26,16 @@ import java.util.List;
  */
 
 public class CityWeatherViewModel extends ViewModel {
-    private Activity activity;
+    private Activity requireActivity;
 
-    public Activity getActivity() {
-        return activity;
+    MutableLiveData<SevenDayWeatherData.DailyDTO> sevenDayWeatherLiveData = new MutableLiveData<>();
+
+    public Activity setRequireActivity() {
+        return requireActivity;
     }
 
-    public void setActivity(Activity activity) {
-        this.activity = activity;
+    public void setRequireActivity(Activity activity) {
+        this.requireActivity = activity;
     }
 
     /**
@@ -40,14 +45,13 @@ public class CityWeatherViewModel extends ViewModel {
      */
     public void getHourlyDTo(String Location, DataCallback<HourlyWeatherData.HourlyDTO> callback) {
 
-        List<HourlyWeatherData.HourlyDTO> mList;
 
         //获取天气的地址
         String address = "https://devapi.qweather.com/v7/weather/24h?location=" + Location + "&key=64f323b501dc410cb7ec4fd1b503aab4";
 
-        Log.e("haojinhui", "getHourlyDTo: " + address);
 
         HourlyWeatherUtility utility = new HourlyWeatherUtility();
+        LogUtil.d("sunwenyu", "dataList.toString()1");
 
         utility.sendAddress(address, new DataCallback<HourlyWeatherData.HourlyDTO>() {
 
@@ -55,27 +59,64 @@ public class CityWeatherViewModel extends ViewModel {
             public void onSuccess(List<HourlyWeatherData.HourlyDTO> dataList) {
 
                 SplitList(dataList);
+
                 callback.onSuccess(dataList);
+            }
+
+            @Override
+            public void onFailure(IOException e) {
+                callback.onFailure(e);
+            }
+
+        });
+    }
+
+    /**
+     * 使用Room保存24小时的天气情况
+     *
+     * @param dataList
+     */
+    public void SplitList(List<HourlyWeatherData.HourlyDTO> dataList) {
+        LogUtil.d("sunwenyu", "dataList.toString()2");
+
+        if (TwentyFourHourWeatherDatabase.Getinstance() != null) {
+            LogUtil.d("sunwenyu", "dataList.toString()3");
+            WeatherDataInquireTool.dpHourWeatherDatabase.weatherDataDao().deleteAll();
+
+            for (HourlyWeatherData.HourlyDTO dto : dataList) {
+//            TwentyFourHourWeatherDataModel weatherData = new TwentyFourHourWeatherDataModel();
+//            weatherData.data = dto;
+//            WeatherDataInquireTool.dpHourWeatherDatabase.weatherDataDao().insertData(weatherData);
+                LogUtil.d("sunwenyu", dto.toString());
+            }
+        }else {
+            LogUtil.d("sunwenyu", dataList.toString());
+        }
+    }
+
+    /**
+     * 获取7日内的天气情况
+     *
+     * @param Location
+     */
+    public void getSevenDayWeather(String Location) {
+        //获取天气的地址
+        String address = "https://devapi.qweather.com/v7/weather/24h?location=" + Location + "&key=64f323b501dc410cb7ec4fd1b503aab4";
+
+
+        SevenDayWeatherUtility utility = new SevenDayWeatherUtility();
+
+        utility.sendAddress(address, new DataCallback() {
+            @Override
+            public void onSuccess(List dataList) {
+
+                sevenDayWeatherLiveData.postValue((SevenDayWeatherData.DailyDTO) dataList);
             }
 
             @Override
             public void onFailure(IOException e) {
 
             }
-
         });
-    }
-
-    public void SplitList(List<HourlyWeatherData.HourlyDTO> dataList){
-        WeatherDataInquireTool.dpHourWeatherDatabase.weatherDataDao().deleteAll();
-
-        for(HourlyWeatherData.HourlyDTO dto : dataList){
-            TwentyFourHourWeatherData weatherData = new TwentyFourHourWeatherData();
-            weatherData.data = dto;
-            LogUtil.d("haojinhui",dto.toString());
-
-            WeatherDataInquireTool.dpHourWeatherDatabase.weatherDataDao().insertData(weatherData);
-        }
-
     }
 }
